python_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
groq==0.4.2
python-dotenv==1.0.0
pydantic==2.5.2
python-docx==0.8.11
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
nltk==3.8.1
spacy==3.7.2
langdetect==1.0.9
tqdm==4.66.1
chardet==5.2.0
python-magic==0.4.27  # For better file type detection
